+++ CHANGELOG.md
@@ -9,7 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Fixed
 
-- Import KML placemarks with a name or description that contains CDATA
+- [Fix KML import crash on placemarks containing CDATA #1100](https://github.com/farmOS/farmOS/pull/1100)
 
 ## [4.0.5] 2026-08-26